inference_service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0